clarify behavior when adding table with and without column mapping
